% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{calc_disponibilidade_estritamente_crescente}
\alias{calc_disponibilidade_estritamente_crescente}
\title{calc_disponibilidade_estritamente_crescente}
\usage{
calc_disponibilidade_estritamente_crescente(disponibilidade_liquida,
  NO_DIA_COMPLETO_dmy)
}
\description{
calc_disponibilidade_estritamente_crescente
}
