% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{calc_integral_sobre_media_dos_gastos}
\alias{calc_integral_sobre_media_dos_gastos}
\title{calc_integral_sobre_media_dos_gastos}
\usage{
calc_integral_sobre_media_dos_gastos(disponibilidade_liquida,
  pagamento_diario)
}
\description{
calc_integral_sobre_media_dos_gastos
}
