% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{calc_valor_nominal_conservador}
\alias{calc_valor_nominal_conservador}
\title{calc_valor_nominal_conservador}
\usage{
calc_valor_nominal_conservador(disponibilidade_liquida, pagamentos_diarios)
}
\description{
calc_valor_nominal_conservador
}
