% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apps.R
\name{explorar}
\alias{explorar}
\title{Explorador dos Perfis de Fluxo de Caixa das UGs}
\usage{
explorar(dados_preparados_mas_sem_indicadores)
}
\arguments{
\item{dados_com_indicadores}{}
}
\description{
Explorador dos Perfis de Fluxo de Caixa das UGs
}
