% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{gera_disponibilidades_liquidas_diarias}
\alias{gera_disponibilidades_liquidas_diarias}
\title{gera_disponibilidades_liquidas_diarias}
\usage{
gera_disponibilidades_liquidas_diarias(lista_dos_tres_itens)
}
\arguments{
\item{lista_dos_tres_itens}{}
}
\description{
gera_disponibilidades_liquidas_diarias
}
