% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{gera_series_diarias_para_os_tres_itens}
\alias{gera_series_diarias_para_os_tres_itens}
\title{gera_series_diarias_para_os_tres_itens}
\usage{
gera_series_diarias_para_os_tres_itens(dados)
}
\description{
gera_series_diarias_para_os_tres_itens
}
