% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{gera_tabela_aninhada}
\alias{gera_tabela_aninhada}
\title{gera_tabela_aninhada}
\usage{
gera_tabela_aninhada(dados_com_disponibilidade_liquida)
}
\arguments{
\item{dados_com_disponibilidade_liquida}{}
}
\description{
gera_tabela_aninhada
}
