% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modelo_suspeita_de_empocamento_v01}
\alias{modelo_suspeita_de_empocamento_v01}
\title{Modelo de Suspeita de Empoçamento}
\format{
An object of class \code{workflow} of length 4.
}
\usage{
modelo_suspeita_de_empocamento_v01
}
\description{
Modelo de Suspeita de Empoçamento
}
\keyword{datasets}
