% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{prepara}
\alias{prepara}
\title{Prepara Perfis Históricos de Fluxo de Caixa}
\usage{
prepara(dados_limsaque_pagto_obrig)
}
\arguments{
\item{dados_limsaque_pagto_obrig}{data.frame extraído do SIAF com os três itens de limite de saque, obricações a pagar e pagamentos.}
}
\value{
data.frame com 5 colunas de identificação e 1 coluna de data.frames aninhados:

\item{id}{}
\item{NO_ORGAO_MAXI}{}
\item{NO_UG}{}
\item{NO_ORGAO}{}
\item{NO_FONTE_RECURSO}{}
\item{series_temporais}{}

data.frame aninhado em series_temporais:

\item{ID_ANO_LANC}{}
\item{NO_DIA_COMPLETO}{}
\item{saldo_diario}{}
\item{flag_saldo_anual}{}
\item{NO_DIA_COMPLETO_dmy}{}
\item{saldo_diario_acumulado}{}
\item{paded}{}
\item{obrigacoes_a_pagar_diario}{}
\item{obrigacoes_a_pagar_diario_acumulado}{}
\item{pagamento_diario}{}
\item{pagamento_diario_acumulado}{}
\item{disponibilidade_liquida}{}
}
\description{
Prepara Perfis Históricos de Fluxo de Caixa
}
\examples{
library(empocamento)
series_mj <- prepara(movimentacoes_diarias_mj)
}
