% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicadores.R
\name{adiciona_descritores}
\alias{adiciona_descritores}
\title{Adiciona Descritores de Perfis de Fluxo de Caixa}
\usage{
adiciona_descritores(dados_aninhados, n_dias = 365)
}
\arguments{
\item{n_dias}{integer. Número de dias de histórico para considerar no cálculo dos descritores. O padrão é 365 dias.}
}
\description{
Adiciona Descritores de Perfis de Fluxo de Caixa
}
\examples{
library(empocamento)
series_mj <- movimentacoes_diarias_mj \%>\%
  prepara() \%>\%
  adiciona_descritores()
}
